requirements.txt
flask
flask-cors
google-genai
requests
python-dotenv
